+++ tests/arrays.rs
@@ -9,6 +9,7 @@ struct AdditionalSettings {
 #[derive(Debug, Default, Miniconf, Deserialize)]
 struct Settings {
     data: u32,
+    #[miniconf(defer)]
     more: AdditionalSettings,
 }
 
@@ -16,6 +17,7 @@ struct Settings {
 fn simple_array() {
     #[derive(Miniconf, Default)]
     struct S {
+        #[miniconf(defer)]
         a: [u8; 3],
     }
 
@@ -39,6 +41,7 @@ fn simple_array() {
 fn nonexistent_field() {
     #[derive(Miniconf, Default)]
     struct S {
+        #[miniconf(defer)]
         a: [u8; 3],
     }
 
@@ -53,6 +56,7 @@ fn nonexistent_field() {
 fn simple_array_indexing() {
     #[derive(Miniconf, Default)]
     struct S {
+        #[miniconf(defer)]
         a: [u8; 3],
     }
 
@@ -73,20 +77,21 @@ fn simple_array_indexing() {
 
     // Test metadata
     let metadata = s.get_metadata();
-    assert_eq!(metadata.max_depth, 3);
+    assert_eq!(metadata.max_depth, 2);
     assert_eq!(metadata.max_topic_size, "a/2".len());
 }
 
 #[test]
 fn array_of_structs_indexing() {
-    #[derive(Miniconf, Default, Clone, Copy, Deserialize, Debug, PartialEq)]
+    #[derive(Miniconf, Default, Clone, Copy, Debug, PartialEq)]
     struct Inner {
         b: u8,
     }
 
     #[derive(Miniconf, Default, PartialEq, Debug)]
     struct S {
-        a: [Inner; 3],
+        #[miniconf(defer)]
+        a: miniconf::Array<Inner, 3>,
     }
 
     let mut s = S::default();
@@ -107,4 +112,48 @@ fn array_of_structs_indexing() {
     let metadata = s.get_metadata();
     assert_eq!(metadata.max_depth, 4);
     assert_eq!(metadata.max_topic_size, "a/2/b".len());
+}
+
+#[test]
+fn array_of_arrays() {
+    #[derive(Miniconf, Default, PartialEq, Debug)]
+    struct S {
+        #[miniconf(defer)]
+        data: miniconf::Array<[u32; 2], 2>,
+    }
+
+    let mut s = S::default();
+
+    let field = "data/0/0".split('/').peekable();
+    s.string_set(field, "7".as_bytes()).unwrap();
+
+    let expected = {
+        let mut e = S::default();
+        e.data[0][0] = 7;
+        e
+    };
+
+    assert_eq!(expected, s);
+}
+
+#[test]
+fn atomic_array() {
+    #[derive(Miniconf, Default, PartialEq, Debug)]
+    struct S {
+        data: [u32; 2],
+    }
+
+    let mut s = S::default();
+
+    let field = "data".split('/').peekable();
+    s.string_set(field, "[1, 2]".as_bytes()).unwrap();
+
+    let expected = {
+        let mut e = S::default();
+        e.data[0] = 1;
+        e.data[1] = 2;
+        e
+    };
+
+    assert_eq!(expected, s);
 }