+++ tests/generics.rs
@@ -1,10 +1,10 @@
-use miniconf::{Miniconf, MiniconfAtomic};
+use miniconf::Miniconf;
 use serde::{Deserialize, Serialize};
 
 #[test]
 fn generic_type() {
     #[derive(Miniconf, Default)]
-    struct Settings<T: Miniconf> {
+    struct Settings<T> {
         pub data: T,
     }
 
@@ -23,7 +23,8 @@ fn generic_type() {
 #[test]
 fn generic_array() {
     #[derive(Miniconf, Default)]
-    struct Settings<T: Miniconf> {
+    struct Settings<T> {
+        #[miniconf(defer)]
         pub data: [T; 2],
     }
 
@@ -36,7 +37,7 @@ fn generic_array() {
 
     // Test metadata
     let metadata = settings.get_metadata();
-    assert_eq!(metadata.max_depth, 3);
+    assert_eq!(metadata.max_depth, 2);
     assert_eq!(metadata.max_topic_size, "data/0".len());
 }
 
@@ -47,7 +48,7 @@ fn generic_struct() {
         pub inner: T,
     }
 
-    #[derive(MiniconfAtomic, Deserialize, Serialize, Default)]
+    #[derive(Serialize, Deserialize, Default)]
     struct Inner {
         pub data: f32,
     }
@@ -72,7 +73,7 @@ fn generic_atomic() {
         pub atomic: Inner<T>,
     }
 
-    #[derive(Deserialize, Serialize, MiniconfAtomic, Default)]
+    #[derive(Deserialize, Serialize, Default)]
     struct Inner<T> {
         pub inner: [T; 5],
     }