+++ CHANGELOG.md
@@ -10,11 +10,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * python module: don't emite whitespace in JSON to match serde-json-core (#92)
 * `heapless::String` now implements `Miniconf` directly.
 * Python client API is no longer retain by default. CLI is unchanged
+* [breaking] Support for `#[derive(MiniconfAtomic)]` was removed.
+* Fields in `#[derive(Miniconf)]` are now atomic by default. To recurse, users must
+  annotate fields with `#[miniconf(defer)]`
+* New `miniconf::Option` type has been added. Existing `Option` implementation has been changed to
+  allow run-time nullability of settings for more flexibility.
+* New `miniconf::Array` type has been added, replacing the previous [T; N] implementation
+* `Miniconf` implementation on most primitive types has been removed as it is no longer required.
 
 ### Fixed
 * Python device discovery now only discovers unique device identifiers. See [#97](https://github.com/quartiq/miniconf/issues/97)
 * Python requests API updated to use a static response topic
 * Python requests now have a timeout
+* Generic trait bound requirements have been made more specific.
 
 
 ## [0.5.0] - 2022-05-12